创建云涂包裹订单

+++ oms/src/main/java/com/upedge/oms/modules/order/vo/AppOrderVo.java
@@ -111,7 +111,7 @@ public class AppOrderVo {
 
     private Integer packState;
 
-    private Integer packNo;
+    private Long packNo;
 
     private Integer stockState;
